pruebas unitarias casi y control de errores casi clase usuarios ya completada

+++ src/main/java/com/example/Services/UserService.java
@@ -1,6 +1,5 @@
 package com.example.Services;
 
-
 import com.example.Entities.User;
 import com.example.Dao.UserDAO;
 import com.example.Errors.ResourceNotFoundException;
@@ -8,27 +7,30 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-
 import java.util.List;
 
 @Service
 public class UserService {
     private static Logger logger = Logger.getLogger(UserService.class);
-    UserDAO userDAO;
+    private final UserDAO userDAO;
+
     @Autowired
-    public UserService(UserDAO userDAO){
+    public UserService(UserDAO userDAO) {
         this.userDAO = userDAO;
     }
 
-    public List<User> readAllUsers(){
+    public List<User> readAllUsers() {
         return userDAO.findAll();
     }
+
     public User createUser(User user) {
         return userDAO.save(user);
     }
 
     public void deleteUser(Integer id) {
-        userDAO.findById(id).orElseThrow(() -> new ResourceNotFoundException("No se puede borrar, porque no existe ningun usuario con esa id."));
+        if (!userDAO.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede borrar, porque no existe ningún usuario con esa ID: " + id);
+        }
         userDAO.deleteById(id);
     }
 
@@ -36,18 +38,17 @@ public class UserService {
         User existingUser = userDAO.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("No se ha encontrado al usuario con el id: " + id));
 
+        existingUser.setContraseña(updatedUser.getContraseña());
+        existingUser.setNombre(updatedUser.getNombre());
+        existingUser.setApellidos(updatedUser.getApellidos());
 
-        existingUser.setEmail(updatedUser.getEmail());
-        existingUser.setPassword(updatedUser.getPassword());
-        existingUser.setNom(updatedUser.getNom());
-        existingUser.setApellido(updatedUser.getApellido());
-        existingUser.setTelefono(updatedUser.getTelefono());
 
         return userDAO.save(existingUser);
     }
+
+
     public User readUserById(Integer id) {
-        return userDAO.findById(id).orElseThrow(() -> new ResourceNotFoundException("No se ha encontrado al usuario con el id: " + id));
-
+        return userDAO.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("No se ha encontrado al usuario con el ID: " + id));
     }
-
 }
